+++ game/src/headers/board.h
@@ -2,11 +2,11 @@
 #define BOARD_H
 #include "globals.h"
 
-int writeBlocks(int x, int y, int block[][SHAPE_SIZE], int board[][ROWS], int flashPiece);
-int colliding(int x, int y, int block[][SHAPE_SIZE], int board[][ROWS]);
-void clearBoard(int board[][ROWS]);
-int checkLineClears(int board[][ROWS], int rowToClear[4]);
-void toggleRowColor(int board[][ROWS], int row, int color);
-void clearLineRows(int board[][ROWS],int rowsToClear[4]);
+int writeBlocks(int x, int y, uint8_t block[][SHAPE_SIZE], uint8_t board[][ROWS], int flashPiece);
+int colliding(int x, int y, uint8_t block[][SHAPE_SIZE], uint8_t board[][ROWS]);
+void clearBoard(uint8_t board[][ROWS]);
+int checkLineClears(uint8_t board[][ROWS], int rowToClear[4]);
+void toggleRowColor(uint8_t board[][ROWS], int row, int color);
+void clearLineRows(uint8_t board[][ROWS], int rowsToClear[4]);
 
 #endif
